+++ backend/src/controllers/auth.controller.ts
@@ -68,45 +68,3 @@ export const loginUser = async (req: Request, res: Response): Promise<void> => {
     res.status(500).json({ message: 'Something went wrong' });
   }
 };
-
-export const updateUserRole = async (req: Request, res: Response) => {
-  try {
-    // 1️⃣ Get user ID from request parameters
-    const { userId, role } = req.body;
-    // 2️⃣ Check if the requestor is a superadmin
-    const requestor = await User.findById(req.userId);
-    if (requestor?.role !== 'superadmin') {
-      return res
-        .status(403)
-        .json({ message: 'Only superadmins can modify roles' });
-    }
-    // 3️⃣ Find the user by ID
-    const user = await User.findByIdAndUpdate(
-      userId,
-      { role },
-      { new: true } // Return the updated user
-    ).select('-password'); // Exclude password from the response
-    // 4️⃣ Check if user was found
-    if (!user) {
-      return res.status(404).json({ message: 'User not found' });
-    }
-    res.status(200).json({ message: 'User role updated successfully', user });
-  } catch (error) {
-    console.log(error);
-    res.status(500).json({ message: 'Error updating user role' });
-  }
-};
-
-export const getAdmins = async (req: Request, res: Response) => {
-  try {
-    // 1️⃣ Get all admins from the database
-    const admins = await User.find({
-      role: { $in: ['admin', 'superadmin'] },
-    }).select('-password');
-    // 2️⃣ Return the list of admins
-    res.status(200).json(admins);
-  } catch (error) {
-    console.log(error);
-    res.status(500).json({ message: 'Error fetching admins' });
-  }
-};
