+++ backend/src/routes/users.route.ts
@@ -1,5 +1,7 @@
 import express from 'express';
 import { registerUser } from '../controllers/user.controller';
+import { verifyToken } from '../middleware/auth';
+import { updateUserRole, getAdmins } from '../controllers/user.controller';
 import { check } from 'express-validator';
 
 //Create a new router instance
@@ -17,5 +19,9 @@ router.post(
   ],
   registerUser
 );
+
+// Define the route for updating user roles
+router.put('/update-role', verifyToken, updateUserRole);
+router.get('/admins', verifyToken, getAdmins);
 
 export default router;
